requirements.txt
scikit-learn
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

hydra-core
hydra_colorlog
mlflow
torch
pytorch-lightning
torchvision
scipy
baal